+++ src/users/users.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
@@ -10,6 +14,15 @@ export class UsersService {
   constructor(private prisma: PrismaService) {}
 
   async create(createUserDto: CreateUserDto) {
+    const existingUser = await this.prisma.user.findUnique({
+      where: { email: createUserDto.email },
+    });
+    if (existingUser) {
+      throw new ConflictException({
+        message: 'Conflict Operation',
+        error: `Article with title ${createUserDto.email} already exist`,
+      });
+    }
     const hashedPassword = await bcrypt.hash(
       createUserDto.password,
       roundsOfHashing,
@@ -23,11 +36,25 @@ export class UsersService {
     return this.prisma.user.findMany();
   }
 
-  findOne(id: number) {
-    return this.prisma.user.findUnique({ where: { id } });
+  async findOne(id: number) {
+    const user = await this.prisma.user.findUnique({ where: { id } });
+    if (!user) {
+      return new NotFoundException({
+        message: 'Not Found',
+        error: `User with id ${id} does not exist`,
+      });
+    }
+    return user;
   }
 
   async update(id: number, updateUserDto: UpdateUserDto) {
+    const user = await this.prisma.user.findUnique({ where: { id } });
+    if (!user) {
+      return new NotFoundException({
+        message: 'Not Found',
+        error: `User with id ${id} does not exist`,
+      });
+    }
     if (updateUserDto.password) {
       const hashedPassword = await bcrypt.hash(
         updateUserDto.password,
@@ -39,7 +66,14 @@ export class UsersService {
     return this.prisma.user.update({ where: { id }, data: updateUserDto });
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const user = await this.prisma.user.findUnique({ where: { id } });
+    if (!user) {
+      return new NotFoundException({
+        message: 'Not Found',
+        error: `User with id ${id} does not exist`,
+      });
+    }
     return this.prisma.user.delete({ where: { id } });
   }
 }
